Внёс мелкие правки в Entity, для обеспечения создания сущностей в БД. Сделал маппинг для моделей и DTO.

+++ src/main/java/pro/sky/graduate_work_group5_team1/Repositories/AdsRepository.java
@@ -2,7 +2,9 @@ package pro.sky.graduate_work_group5_team1.Repositories;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import pro.sky.graduate_work_group5_team1.entity.Ads;
+import pro.sky.graduate_work_group5_team1.model.Ads;
+
 @Repository
-public interface AdsRepository extends JpaRepository <Ads, Long> {
+public interface AdsRepository extends JpaRepository<Ads, Integer> {
+
 }
